+++ exo-go/src/application/service_restarter.go
@@ -1,4 +1,4 @@
-package serviceRestarter
+package application
 
 import (
 	"fmt"
@@ -9,7 +9,7 @@ import (
 )
 
 // ServiceRestarter watches the given local service for changes and restarts it
-type ServiceRestarter struct {
+type serviceRestarter struct {
 	ServiceName      string
 	ServiceDir       string
 	DockerComposeDir string
@@ -19,7 +19,7 @@ type ServiceRestarter struct {
 }
 
 // Watch watches the service directory for changes
-func (s *ServiceRestarter) Watch(watcherErrChannel chan<- error) {
+func (s *serviceRestarter) Watch(watcherErrChannel chan<- error) {
 	watcher, err := fsnotify.NewWatcher()
 	if err != nil {
 		watcherErrChannel <- err
@@ -50,19 +50,7 @@ func (s *ServiceRestarter) Watch(watcherErrChannel chan<- error) {
 	}
 }
 
-func isChange(event fsnotify.Event) bool {
-	return (event.Op&fsnotify.Write == fsnotify.Write) || (event.Op&fsnotify.Rename == fsnotify.Rename) || (event.Op&fsnotify.Chmod == fsnotify.Chmod)
-}
-
-func isCreate(event fsnotify.Event) bool {
-	return event.Op&fsnotify.Create == fsnotify.Create
-}
-
-func isRemove(event fsnotify.Event) bool {
-	return event.Op&fsnotify.Remove == fsnotify.Remove
-}
-
-func (s *ServiceRestarter) restart(watcherErrChannel chan<- error) {
+func (s *serviceRestarter) restart(watcherErrChannel chan<- error) {
 	if err := s.watcher.Close(); err != nil {
 		watcherErrChannel <- err
 		return
@@ -83,4 +71,18 @@ func (s *ServiceRestarter) restart(watcherErrChannel chan<- error) {
 	}
 	s.LogChannel <- fmt.Sprintf("'%s' restarted successfully", s.ServiceName)
 	s.Watch(watcherErrChannel)
+}
+
+// Helpers
+
+func isChange(event fsnotify.Event) bool {
+	return (event.Op&fsnotify.Write == fsnotify.Write) || (event.Op&fsnotify.Rename == fsnotify.Rename) || (event.Op&fsnotify.Chmod == fsnotify.Chmod)
+}
+
+func isCreate(event fsnotify.Event) bool {
+	return event.Op&fsnotify.Create == fsnotify.Create
+}
+
+func isRemove(event fsnotify.Event) bool {
+	return event.Op&fsnotify.Remove == fsnotify.Remove
 }
